requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
discord.py==1.3.4
face-alignment
ffmpeg==1.4
future==0.18.2
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.0
multidict==4.7.6
networkx==2.4
numpy==1.18.5
opencv-python==4.3.0.36
pandas==1.1.0
Pillow==7.2.0
pycparser==2.20
PyNaCl==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2020.7.24
--find-links https://download.pytorch.org/whl/torch_stable.html torch==1.6.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html torchvision==0.7.0
tqdm==4.24.0
typing-extensions==3.7.4.2
websockets==8.1
wincertstore==0.2
yarl==1.5.0
youtube-dl==2020.7.28
